+++ src/main/java/cz/cvut/fel/constructa/controller/UserControllerImpl.java
@@ -4,14 +4,20 @@ import cz.cvut.fel.constructa.controller.interfaces.UserController;
 import cz.cvut.fel.constructa.dto.request.UserRequest;
 import cz.cvut.fel.constructa.dto.response.UserDTO;
 import cz.cvut.fel.constructa.dto.response.UserInputDTO;
+import cz.cvut.fel.constructa.security.AuthenticationFacade;
 import cz.cvut.fel.constructa.service.interfaces.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.web.bind.annotation.*;
 
 import java.text.ParseException;
+import java.util.Arrays;
+import java.util.Collection;
 import java.util.List;
 
 /**
@@ -23,9 +29,26 @@ import java.util.List;
 @RequiredArgsConstructor
 public class UserControllerImpl implements UserController {
     /**
+     * The Authentication facade.
+     */
+    private final AuthenticationFacade authenticationFacade;
+
+    /**
      * The User service.
      */
     private final UserService userService;
+
+    /**
+     * Has permission boolean.
+     *
+     * @param requiredRoles the required roles
+     * @return the boolean
+     */
+    private boolean hasPermission(List<GrantedAuthority> requiredRoles){
+        Authentication authentication = authenticationFacade.getAuthentication();
+        Collection<? extends GrantedAuthority> authorities = authentication.getAuthorities();
+        return authorities.stream().anyMatch(requiredRoles::contains);
+    }
 
     /**
      * Gets users.
@@ -88,8 +111,18 @@ public class UserControllerImpl implements UserController {
     @ResponseBody
     @PutMapping(produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<UserDTO> updateUser(@RequestBody UserRequest request) throws ParseException {
-        return ResponseEntity.ok().body(
-                userService.update(request));
+        List<GrantedAuthority> requiredRoles = Arrays.asList(
+                new SimpleGrantedAuthority("ROLE_ADMIN"),
+                new SimpleGrantedAuthority("ROLE_MANAGER")
+        );
+
+        if (hasPermission(requiredRoles)) {
+            return ResponseEntity.ok().body(
+                    userService.update(request)
+            );
+        } else {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
     }
 
     /**
@@ -102,7 +135,17 @@ public class UserControllerImpl implements UserController {
     @ResponseStatus(code = HttpStatus.NO_CONTENT)
     @DeleteMapping(value = "/{userId}")
     public ResponseEntity<Void> deleteUser(@PathVariable Long userId) {
-        userService.delete(userId);
-        return ResponseEntity.noContent().build();
+        List<GrantedAuthority> requiredRoles = Arrays.asList(
+                new SimpleGrantedAuthority("ROLE_ADMIN"),
+                new SimpleGrantedAuthority("ROLE_MANAGER")
+        );
+
+        if (hasPermission(requiredRoles)) {
+            userService.delete(userId);
+            return ResponseEntity.noContent().build();
+        } else {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
     }
 }
